+++ src/game/resources/abilities/Drain.java
@@ -21,7 +21,7 @@ public class Drain extends Ability {
     }
 
     @Override
-    public float getDeflectionDamage(Hero other, int receivedRawDamage) {
+    public float getDeflectionDamage(Hero other, float receivedRawDamage) {
         return 0;
     }
