+++ src/game/resources/abilities/Fireblast.java
@@ -24,7 +24,7 @@ public class Fireblast extends Ability {
     }
 
     @Override
-    public float getDeflectionDamage(Hero other, int receivedRawDamage) {
+    public float getDeflectionDamage(Hero other, float receivedRawDamage) {
         return 0;
     }
