+++ src/game/resources/abilities/Paralysis.java
@@ -29,7 +29,7 @@ public class Paralysis extends Ability {
     }
 
     @Override
-    public float getDeflectionDamage(Hero other, int receivedRawDamage) {
+    public float getDeflectionDamage(Hero other, float receivedRawDamage) {
         return 0;
     }
